EthChainController renamed with EVMNetworkController
