+++ source/scripts/Background/controllers/EVMNetworkController/index.ts
@@ -1,4 +1,4 @@
-export { default } from './EthChainController';
+export { default } from './EVMNetworkController';
 export * from './AccountTracker';
 import * as utils from '@xchainjs/xchain-util';
 export { utils };
