+++ source/scripts/Background/controllers/EVMNetworkController/utils.ts
