+++ source/state/nfts/types.ts
@@ -1,4 +1,4 @@
-import { EthNetworkId } from 'scripts/Background/controllers/EthChainController/types';
+import { EthNetworkId } from 'scripts/Background/controllers/EVMNetworkController/types';
 import { AssetType } from 'state/vault/types';
 
 export interface INFTInfoState {
