+++ source/state/vault/types.ts
@@ -6,7 +6,7 @@ import {
   KeyringWalletState,
   KeyringWalletType,
 } from '@stardust-collective/dag4-keyring';
-import { EthNetworkId } from 'scripts/Background/controllers/EthChainController/types';
+import { EthNetworkId } from 'scripts/Background/controllers/EVMNetworkController/types';
 
 export type SeedKeystore = V3Keystore<KDFParamsPhrase>;
 export type PrivKeystore = V3Keystore<KDFParamsPrivateKey>;
